expose generate_alignment
