+++ palamedes/__init__.py
@@ -1,4 +1,4 @@
-from Bio.Align import PairwiseAligner
+from Bio.Align import PairwiseAligner, Alignment
 from Bio.SeqRecord import SeqRecord
 from hgvs.sequencevariant import SequenceVariant
 
@@ -9,6 +9,118 @@ from palamedes.hgvs.builders import BUILDER_CONFIG
 
 
 __version__ = "0.0.8"
+
+
+def generate_hgvs_variants_from_alignment(
+    alignment: Alignment, use_non_standard_substitution_rules: bool = False, molecule_type: str = MOLECULE_TYPE_PROTEIN
+) -> list[SequenceVariant]:
+    variant_blocks = generate_variant_blocks(
+        alignment,
+        split_consecutive_mismatches=use_non_standard_substitution_rules,
+    )
+    builder = BUILDER_CONFIG[molecule_type](alignment)
+    return [
+        builder.build(
+            variant_block,
+            categorize_variant_block(variant_block, alignment),
+        )
+        for variant_block in variant_blocks
+    ]
+
+
+def generate_alignment(
+    reference_seq_record: SeqRecord,
+    alternate_seq_record: SeqRecord,
+    molecule_type: str = MOLECULE_TYPE_PROTEIN,
+    aligner: PairwiseAligner | None = None,
+) -> Alignment:
+    """
+    Using biopython's PairwiseAligner, generate an alignment object representing the best alignment
+    between the 2 biopython SeqRecords. Note that the molecule_type argument will be used to specify a molecule_type
+    annotation on the SeqRecord.
+
+    By default the function creates an aligner object using the defaults, but the caller may provide their
+    own pre-configured aligner. This aligner must be set to 'global' mode.
+
+    Note that it is possible for multiple alignments to be returned with the same max score. Synthetic testing
+    has shown that generally the highest scoring alignments are returned in "left to right" order. Take the
+    following example, which was for testing HGVS duplications. This was the "intended" alignment:
+
+        ATC---GGGGGGGG
+        ATCATCGGGGGGGG
+
+    However, there are multiple ways to get a three base insertion, all with the same score:
+        ipdb> for aln in aligner.align('ATCGGGGGGGG', 'ATCATCGGGGGGGG'):
+        print(aln)
+
+        target            0 ---ATCGGGGGGGG 11
+                          0 ---||||||||||| 14
+        query             0 ATCATCGGGGGGGG 14
+
+        target            0 A---TCGGGGGGGG 11
+                          0 |---|||||||||| 14
+        query             0 ATCATCGGGGGGGG 14
+
+        target            0 AT---CGGGGGGGG 11
+                          0 ||---||||||||| 14
+        query             0 ATCATCGGGGGGGG 14
+
+        target            0 ATC---GGGGGGGG 11
+                          0 |||---|||||||| 14
+        query             0 ATCATCGGGGGGGG 14
+
+    Due to the HGVS rule of always indicating that the 3' most base is considered the modified one, a best effort
+    attempt is made to return the most "right aligned" alignment, by returning the last alignment with the highest
+    score. This way not yield the ideal results in more complicated cases.
+    """
+    if aligner is not None:
+        if aligner.mode != GLOBAL_ALIGN_MODE:
+            raise ValueError(f"Custom PairwiseAligner must be set to global mode, got: {aligner.mode}")
+    else:
+        aligner = PairwiseAligner(
+            mode=GLOBAL_ALIGN_MODE,
+            match_score=DEFAULT_MATCH_SCORE,
+            mismatch_score=DEFAULT_MISMATCH_SCORE,
+            open_gap_score=DEFAULT_OPEN_GAP_SCORE,
+            extend_gap_score=DEFAULT_EXTEND_GAP_SCORE,
+        )
+
+    if (ref_molecule_type := reference_seq_record.annotations.get("molecule_type")) != molecule_type:
+        raise ValueError(
+            "Cannot generate alignment, reference_seq_record is a SeqRecord an invalid molecule_type annotation "
+            f"got: {ref_molecule_type}, expected: {molecule_type}!"
+        )
+
+    if (alt_molecule_type := alternate_seq_record.annotations.get("molecule_type")) != molecule_type:
+        raise ValueError(
+            "Cannot generate alignment, alternate_seq_record is a SeqRecord an invalid molecule_type annotation "
+            f"got: {alt_molecule_type}, expected: {molecule_type}!"
+        )
+
+    # reverse the sequences and then align the reversed, keeping the first best alignment
+    reversed_ref_seq_record = reverse_seq_record(reference_seq_record)
+    reversed_alt_seq_record = reverse_seq_record(alternate_seq_record)
+    reversed_alignments = aligner.align(reversed_ref_seq_record, reversed_alt_seq_record)
+    reversed_alignment = reversed_alignments[0]
+
+    # undo the reversal, to recover the "last" highest scoring alignment for the forward
+    # which should correspond to the 3' end most alignment and follow HGSV spec
+    # note 'infer_coordinates' will soon be deprecated, update to 'parse_printed_alignment' once its live
+    forward_coordinates = Alignment.infer_coordinates(
+        [
+            reversed_alignment[0][::-1],
+            reversed_alignment[1][::-1],
+        ]
+    )
+    forward_alignment = Alignment(
+        [reference_seq_record, alternate_seq_record],
+        forward_coordinates,
+    )
+
+    # score is not technically an attribute on the class
+    setattr(forward_alignment, "score", reversed_alignment.score)
+
+    return forward_alignment
 
 
 def generate_hgvs_variants(
@@ -85,15 +197,4 @@ def generate_hgvs_variants(
     )
 
     alignment = generate_alignment(ref_seq_record, alt_seq_record, molecule_type=molecule_type, aligner=aligner)
-    variant_blocks = generate_variant_blocks(
-        alignment,
-        split_consecutive_mismatches=use_non_standard_substitution_rules,
-    )
-    builder = BUILDER_CONFIG[molecule_type](alignment)
-    return [
-        builder.build(
-            variant_block,
-            categorize_variant_block(variant_block, alignment),
-        )
-        for variant_block in variant_blocks
-    ]
+    return generate_hgvs_variants_from_alignment(alignment, use_non_standard_substitution_rules)
